+++ packages/flutter_timeline/example/lib/main.dart
@@ -1,6 +1,6 @@
 // import 'package:example/apps/go_router/app.dart';
-// import 'package:example/apps/navigator/app.dart';
-import 'package:example/apps/widgets/app.dart';
+import 'package:example/apps/navigator/app.dart';
+// import 'package:example/apps/widgets/app.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
@@ -9,7 +9,7 @@ void main() {
 
   // Uncomment any, but only one, of these lines to run the example with specific navigation.
 
-  runApp(const WidgetApp());
-  // runApp(const NavigatorApp());
+  // runApp(const WidgetApp());
+  runApp(const NavigatorApp());
   // runApp(const GoRouterApp());
 }