+++ packages/flutter_timeline_view/lib/src/config/post_config.dart
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2024 Iconica
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+class TimelinePostConfig {
+  const TimelinePostConfig({
+    this.doubleTapTolike = false,
+  });
+
+  /// Whether to allow double tap to like
+  final bool doubleTapTolike;
+}