+++ packages/flutter_timeline_view/lib/src/screens/timeline_selection_screen.dart
@@ -31,9 +31,8 @@ class TimelineSelectionScreen extends StatelessWidget {
             padding: EdgeInsets.only(top: size.height * 0.05, bottom: 8),
             child: Text(
               options.translations.timelineSelectionDescription,
-              style:
-                  options.theme.textStyles.categorySelectionDescriptionStyle ??
-                      theme.textTheme.displayMedium,
+              style: options.textStyles.categorySelectionDescriptionStyle ??
+                  theme.textTheme.displayMedium,
             ),
           ),
           const SizedBox(height: 4),
@@ -55,7 +54,7 @@ class TimelineSelectionScreen extends StatelessWidget {
                 margin: const EdgeInsets.symmetric(vertical: 8),
                 child: Text(
                   category.title,
-                  style: options.theme.textStyles.categorySelectionTitleStyle ??
+                  style: options.textStyles.categorySelectionTitleStyle ??
                       theme.textTheme.displaySmall,
                 ),
               ),